[#250] refactor: 추천사이트 패키지 이동 및 timer service이동
