+++ linkmind/src/main/java/com/app/toaster/timer/controller/TimerController.java
@@ -7,7 +7,7 @@ import com.app.toaster.timer.controller.request.UpdateTimerCommentDto;
 import com.app.toaster.timer.controller.request.UpdateTimerDateTimeDto;
 import com.app.toaster.timer.controller.response.GetTimerResponseDto;
 import com.app.toaster.exception.Success;
-import com.app.toaster.service.timer.TimerService;
+import com.app.toaster.timer.service.TimerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
